% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_starting_tree_v1.R
\name{construct_starting_tree}
\alias{construct_starting_tree}
\title{Construct starting tree}
\usage{
construct_starting_tree(
  OTUs_df = NULL,
  taxa_df = NULL,
  nodes_df = NULL,
  xlsfn = NULL,
  min_brlen = 1e-04,
  outfn = ""
)
}
\arguments{
\item{OTUs_df}{A data.frame with, at least, columns labeled
"OTUs" (tip names),
"use" (which OTUs will be used, "yes" or "no"; blanks etc. mean "yes"), and
"tipdate" (the desired tipdates; use 0, if all the tips are contemporaneous)
If OTUs_df is NULL, it can be read from the Excel file xlsfn; the program
will skip the first 14 lines of the worksheet; the 15th row will have the
column headers.  This is the exact command used by construct_starting_tree():
readWorksheetFromFile(xlsfn, sheet="OTUs", startRow=15)}

\item{taxa_df}{A data.frame with, at least, columns listing the OTUs that
go into each clade. The column headers are the names of the clades.
construct_starting_tree() assumes that one column will have ALL of the
OTUs, and will be named "total_group_LCA". This clade represents the
Last Common Ancestor (LCA) of the total group.

All clade names in taxa_df should match a clade name in nodes_df$Taxon.
Those that don't will be cut.

As clades will have OTU lists of different lengths, the extra spaces in the
table should be blanks ("") or NA.
OTUs missing from OTUs_df, or marked $use="no", will be cut.
If taxa_df is NULL, it can be read from the Excel file xlsfn; the program
will skip the first 14 lines of the worksheet; the 15th row will have the
column headers.  This is the exact command used by construct_starting_tree():
readWorksheetFromFile(xlsfn, sheet="taxa", startRow=15)}

\item{nodes_df}{A data.frame with, at least, columns labeled
"Taxon" (clade names, or non-clade groups of OTUs),
"use" (which OTUs will be used, "yes" or "no"; blanks etc. mean "yes")
"mono" Is the group monophyletic? "yes" or "no". Only "yes" will be used
for building the starting tree.
"make_age_prior" Should be "yes" or "no". Only "yes" will be used
in the starting tree
"distribution" What is the prior distribution on the node age? Can be
one of: normal, uniform, lognormal, exponential
meanInRealSpace For the lognormal distribution, is the mean parameter
(param1) the mean in real space ("yes"), or not ("no"). The function
will convert between these.
offset: Offset to the mean (for lognormal or exponential)
param1: the mean, lower bound, mean, and mean, respectively
param2: the standard deviation, upper bound, standard deviation,
and ignored, respectively
If nodes_df is NULL, it can be read from the Excel file xlsfn; the program
will skip the first 14 lines of the worksheet; the 15th row will have the
column headers.  This is the exact command used by construct_starting_tree():
readWorksheetFromFile(xlsfn, sheet="nodes", startRow=15)}

\item{xlsfn}{If the above inputs are NULL, then OTUs_df, nodes_df, and taxa_df
can be read from an Excel file formated according to the BEASTmasteR
example settings file.}

\item{min_brlen}{During construction of the starting tree, all internal
(non-tip) branchlengths will be min_brlen=0.001 (default). Changing
this will make these larger, but might cause difficulties if you have
node-date constraints that are close together}

\item{outfn}{Write the tree to this file, if not "" (default is "", i.e. blank)}
}
\description{
Construct starting tree
}
