% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_v1.R
\name{findall}
\alias{findall}
\title{Get indices of all matches to a list}
\usage{
findall(what, inlist)
}
\arguments{
\item{what}{The item to find}

\item{inlist}{The list to search in}
}
\value{
\code{matching_indices} List of the matching indices
}
\description{
Just a handy shortcut function
}
\note{
Go BEARS!
}
\examples{
test=1
}
\references{
\url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
}
\seealso{
\code{\link{get_daughters}}, \code{\link{chainsaw2}}
}
\author{
Nicholas J. Matzke \email{matzke@berkeley.edu}
}
