% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_v1.R
\name{get_edge_times_before_present}
\alias{get_edge_times_before_present}
\title{Get the times of the top and bottom of each edge}
\usage{
get_edge_times_before_present(t)
}
\arguments{
\item{t}{An ape phylo object}
}
\value{
\code{edge_times_bp} A 2-column matrix with the age (from the present) of the top and bottom of each edge.
}
\description{
A utility function.
}
\note{
Go BEARS!
}
\examples{
test=1
}
\references{
\url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
}
\seealso{
\code{\link{prt}}, \code{\link{chainsaw2}}
}
\author{
Nicholas J. Matzke \email{matzke@berkeley.edu}
}
