% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_v1.R
\name{is.not.na}
\alias{is.not.na}
\title{Check for not NA}
\usage{
is.not.na(x)
}
\arguments{
\item{x}{Thing to check for NA}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
A utility function.
}
\note{
Go BEARS!
}
\examples{
test=1
}
\references{
\url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
}
\seealso{
\code{\link{prt}}, \code{\link{chainsaw2}}
}
\author{
Nicholas J. Matzke \email{matzke@berkeley.edu}
}
