% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nexus_tree_v1.R
\name{read_nexus2}
\alias{read_nexus2}
\title{Obj The output is a "charslist": a list of species, each with a
list of textual character states. This is the same format as
produced by read_nexus_data2(). A charlist can be converted to a
charsdf (a data.frame) with as.data.frame(x=..., stringsAsFactors=FALSE).
The reverse operation can occur with as.list().}
\usage{
read_nexus2(file, tree.names = NULL, force.multi = FALSE)
}
\description{
Obj The output is a "charslist": a list of species, each with a
list of textual character states. This is the same format as
produced by read_nexus_data2(). A charlist can be converted to a
charsdf (a data.frame) with as.data.frame(x=..., stringsAsFactors=FALSE).
The reverse operation can occur with as.list().
}
