% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_v1.R
\name{sourceall}
\alias{sourceall}
\title{Source all .R files in a directory, except "compile" and "package" files}
\usage{
sourceall(path = path, pattern = "\\\\.R", ...)
}
\arguments{
\item{path}{The path to source}

\item{pattern}{Default is .R}

\item{...}{Additional arguments to source}
}
\value{
\code{path} The path that was sourced.
}
\description{
Utility function.
}
\note{
Go BEARS!
}
\examples{
test=1
}
\references{
\url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
\url{https://code.google.com/p/lagrange/}
}
\seealso{
\code{\link[base]{source}}
}
\author{
Nicholas J. Matzke \email{matzke@berkeley.edu}
}
