% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_v1.R
\name{unlist_dtf_cols}
\alias{unlist_dtf_cols}
\title{Unlist the columns in a data.frame}
\usage{
unlist_dtf_cols(dtf, printflag = FALSE)
}
\arguments{
\item{dtf}{Input \code{\link[base]{data.frame}}}

\item{printflag}{Print the results if TRUE.}
}
\value{
\code{dtf} The data.frame, hopefully without lists for columns
}
\description{
Utility function. What it says.
}
\note{
Go BEARS!
}
\examples{
test=1
}
\references{
\url{http://phylo.wikidot.com/matzke-2013-international-biogeography-society-poster}
\url{https://code.google.com/p/lagrange/}
}
\seealso{
\code{\link[base]{unlist}}
}
\author{
Nicholas J. Matzke \email{matzke@berkeley.edu}
}
