% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nexus_data2_v1.R
\name{write_nexus_data2}
\alias{write_nexus_data2}
\title{x The input is a "charslist": a list of species, each with a
list of textual character states. This is the same format as
produced by read_nexus_data2(). A charlist can be converted to a
charsdf (a data.frame) with as.data.frame(x=..., stringsAsFactors=FALSE).
The reverse operation can occur with as.list().}
\usage{
write_nexus_data2(
  x,
  file,
  format = "dna",
  datablock = TRUE,
  interleaved = TRUE,
  charsperline = NULL,
  gap = NULL,
  missing = NULL
)
}
\description{
x The input is a "charslist": a list of species, each with a
list of textual character states. This is the same format as
produced by read_nexus_data2(). A charlist can be converted to a
charsdf (a data.frame) with as.data.frame(x=..., stringsAsFactors=FALSE).
The reverse operation can occur with as.list().
}
